connecting-to-a-sql-database-project-tutorial-main/requirements.txt
sqlalchemy==1.4.37
pymysql==1.0.2
pandas==1.4.2
python-dotenv==0.20.0